Bug 575436 - rework JavaModelManager.getZipFile()
Moved code from JavaModelManager to new seperate Classes:
*ThreadLocalZipFiles: general caching for ZipFiles
*ZipState: java model handling of invalid ZipFiles

"throwIoExceptionsInGetZipFile" was not moved but deleted to avoid
testcode in production code. Therefore
testTransitionFromInvalidToValidJar() was adapted to produce a real
invalid jar.

ThreadLocalZipFiles contains now AutoCloseable Helper for
try-with-resource resource management.

Calls to the refactored code (in "jdt.core") where refactored to use new
the new try-with-resource pattern.

There is a little difference between the new behavior of old
JavaModelManager.getZipFile() during a JavaModelManager.cacheZipFiles()
block() and new ZipState.createZipFile() in a
ThreadLocalZipFiles.createZipHolder() block:
When blocks are nested they now return the same instance on any nesting
level while previously new instance of ZipFile had been created on every
nesting level.

There is a little difference between the new behavior of old
JavaModelManager.flushZipFiles and new ZipState.close():
Previously flushZipFiles did close all zipFiles in the innermost block;
Now ZipState.close() does only close a zipFile if it was not allocated
in a outer block.

I.e. we have now well defined try-with-resource block: During execution
of any block there is only a single instance of ZipFile for each file.

To use this new behavior builder.ClasspathJar (and its subclasses)
needed to be updated to not hold a local instance of ZipFile (which
could have been leaked).

--
JavaModelManager.getLocalFile(IPath) did return a wrong (relative) path
for deleted workspace resource. Did not matter much as the file did not
exist anyway but it destroyed JavaSearchBugsTests.testBug261722() when
caching filename instead of IPath.

--
In AddJarFileToIndex there is a big block indent that git-diff doesnt
show as such.

Change-Id: Idd94ae90b375db3e475432e710ae7babb99cc5ac